backend/requirements.txt
fastapi
uvicorn
google-generativeai
chromadb
python-multipart
pypdf
python-dotenv
